+++ docs/changelog.rst
@@ -5,6 +5,15 @@
 Changelog
 =========
 
+Version 1.3.0 - Unreleased
+--------------------------
+
+* Moved ``parser.ScssImportsParser`` to ``parser.imports.ScssImportsParser``;
+* Moved comments parsing to its own parser ``parser.comments.ScssCommentsParser``;
+* Make ``parser.imports.ScssImportsParser`` inherit from ``parser.comments.ScssCommentsParser``;
+* Added ``parser.variables.ScssVariablesParser`` that can parse and retrieve every variables;
+* Updated core API documentation;
+
 Version 1.2.2 - 2017/12/12
 --------------------------